Refactor function and class into their own functions
